Separated into internal and pkg
