+++ cmd/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "gitlab.com/gojis/vm"
+import "gitlab.com/gojis/vm/pkg/vm"
 
 func main() {
 	vm.Run()
