+++ internal/runtime/binding/environment_global.go
@@ -3,8 +3,8 @@ package binding
 import (
 	"fmt"
 
-	"gitlab.com/gojis/vm/runtime/errors"
-	"gitlab.com/gojis/vm/runtime/lang"
+	"gitlab.com/gojis/vm/internal/runtime/errors"
+	"gitlab.com/gojis/vm/internal/runtime/lang"
 )
 
 var _ Environment = (*GlobalEnvironment)(nil)
