+++ internal/runtime/lang/type_string_or_symbol.go
